+++ src/LocalResource.js
